# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Architecture Review Process Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface JiraInitiative {
  id: number;
  jiraKey: string;
  summary: string;
  description?: string | null;
  projectKey: string;
  projectName: string;
  status: string;
  priority: string;
  assignee?: string | null;
  issueType: string;
  labels: string[];
  jiraUrl?: string | null;
  syncedAt: string;
  createdAt: string;
}

export interface JiraSyncResult {
  synced: number;
  added: number;
  updated: number;
  lastSyncedAt: string;
}

export type ArchitectureRequestRequestType =
  (typeof ArchitectureRequestRequestType)[keyof typeof ArchitectureRequestRequestType];

export const ArchitectureRequestRequestType = {
  new_technology: "new_technology",
  replacement_migration: "replacement_migration",
  new_capability: "new_capability",
  expansion: "expansion",
  ma_assessment: "ma_assessment",
} as const;

export type ArchitectureRequestPhase =
  (typeof ArchitectureRequestPhase)[keyof typeof ArchitectureRequestPhase];

export const ArchitectureRequestPhase = {
  ph1: "ph1",
  ph2: "ph2",
  ph3: "ph3",
} as const;

export type ArchitectureRequestStatus =
  (typeof ArchitectureRequestStatus)[keyof typeof ArchitectureRequestStatus];

export const ArchitectureRequestStatus = {
  submitted: "submitted",
  ea_triage: "ea_triage",
  specifications_required: "specifications_required",
  arc_scheduled: "arc_scheduled",
  arc_review: "arc_review",
  approved: "approved",
  approved_with_conditions: "approved_with_conditions",
  deferred: "deferred",
  rejected: "rejected",
} as const;

export type ArchitectureRequestPriority =
  (typeof ArchitectureRequestPriority)[keyof typeof ArchitectureRequestPriority];

export const ArchitectureRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface ArchitectureRequest {
  id: number;
  title: string;
  description: string;
  requestType: ArchitectureRequestRequestType;
  phase: ArchitectureRequestPhase;
  submittedBy: string;
  businessUnit: string;
  status: ArchitectureRequestStatus;
  priority: ArchitectureRequestPriority;
  eaAssignee?: string | null;
  architectureSpecifications?: string | null;
  scopeNotes?: string | null;
  jiraInitiativeId?: number | null;
  jiraKey?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateArchitectureRequestRequestType =
  (typeof CreateArchitectureRequestRequestType)[keyof typeof CreateArchitectureRequestRequestType];

export const CreateArchitectureRequestRequestType = {
  new_technology: "new_technology",
  replacement_migration: "replacement_migration",
  new_capability: "new_capability",
  expansion: "expansion",
  ma_assessment: "ma_assessment",
} as const;

export type CreateArchitectureRequestPhase =
  (typeof CreateArchitectureRequestPhase)[keyof typeof CreateArchitectureRequestPhase];

export const CreateArchitectureRequestPhase = {
  ph1: "ph1",
  ph2: "ph2",
  ph3: "ph3",
} as const;

export type CreateArchitectureRequestPriority =
  (typeof CreateArchitectureRequestPriority)[keyof typeof CreateArchitectureRequestPriority];

export const CreateArchitectureRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreateArchitectureRequest {
  title: string;
  description: string;
  requestType: CreateArchitectureRequestRequestType;
  phase: CreateArchitectureRequestPhase;
  submittedBy: string;
  businessUnit: string;
  priority: CreateArchitectureRequestPriority;
  architectureSpecifications?: string | null;
  jiraInitiativeId?: number | null;
}

export type UpdateArchitectureRequestStatus =
  (typeof UpdateArchitectureRequestStatus)[keyof typeof UpdateArchitectureRequestStatus];

export const UpdateArchitectureRequestStatus = {
  submitted: "submitted",
  ea_triage: "ea_triage",
  specifications_required: "specifications_required",
  arc_scheduled: "arc_scheduled",
  arc_review: "arc_review",
  approved: "approved",
  approved_with_conditions: "approved_with_conditions",
  deferred: "deferred",
  rejected: "rejected",
} as const;

export type UpdateArchitectureRequestPriority =
  (typeof UpdateArchitectureRequestPriority)[keyof typeof UpdateArchitectureRequestPriority];

export const UpdateArchitectureRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface UpdateArchitectureRequest {
  status?: UpdateArchitectureRequestStatus;
  eaAssignee?: string | null;
  architectureSpecifications?: string | null;
  scopeNotes?: string | null;
  priority?: UpdateArchitectureRequestPriority;
  jiraInitiativeId?: number | null;
}

export type ArcSessionStatus =
  (typeof ArcSessionStatus)[keyof typeof ArcSessionStatus];

export const ArcSessionStatus = {
  scheduled: "scheduled",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface ArcSession {
  id: number;
  requestId: number;
  scheduledDate: string;
  duration: number;
  status: ArcSessionStatus;
  attendees: string[];
  notes?: string | null;
  createdAt: string;
}

export interface CreateArcSession {
  requestId: number;
  scheduledDate: string;
  duration: number;
  attendees: string[];
  notes?: string | null;
}

export type UpdateArcSessionStatus =
  (typeof UpdateArcSessionStatus)[keyof typeof UpdateArcSessionStatus];

export const UpdateArcSessionStatus = {
  scheduled: "scheduled",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateArcSession {
  status?: UpdateArcSessionStatus;
  notes?: string | null;
  scheduledDate?: string;
}

export type ReviewOutcomeDecision =
  (typeof ReviewOutcomeDecision)[keyof typeof ReviewOutcomeDecision];

export const ReviewOutcomeDecision = {
  approved: "approved",
  approved_with_conditions: "approved_with_conditions",
  deferred: "deferred",
  rejected: "rejected",
  escalated: "escalated",
} as const;

export type ReviewOutcomeOutcomeType =
  (typeof ReviewOutcomeOutcomeType)[keyof typeof ReviewOutcomeOutcomeType];

export const ReviewOutcomeOutcomeType = {
  adr_update: "adr_update",
  deviation: "deviation",
  risk_ownership: "risk_ownership",
  lt_escalation: "lt_escalation",
  standard: "standard",
} as const;

export interface ReviewOutcome {
  id: number;
  requestId: number;
  sessionId?: number | null;
  decision: ReviewOutcomeDecision;
  outcomeType: ReviewOutcomeOutcomeType;
  adrReference?: string | null;
  exceptionOwner?: string | null;
  remediationPlan?: string | null;
  riskOwner?: string | null;
  escalationReason?: string | null;
  nextSteps?: string | null;
  notes?: string | null;
  createdAt: string;
  createdBy: string;
}

export type CreateReviewOutcomeDecision =
  (typeof CreateReviewOutcomeDecision)[keyof typeof CreateReviewOutcomeDecision];

export const CreateReviewOutcomeDecision = {
  approved: "approved",
  approved_with_conditions: "approved_with_conditions",
  deferred: "deferred",
  rejected: "rejected",
  escalated: "escalated",
} as const;

export type CreateReviewOutcomeOutcomeType =
  (typeof CreateReviewOutcomeOutcomeType)[keyof typeof CreateReviewOutcomeOutcomeType];

export const CreateReviewOutcomeOutcomeType = {
  adr_update: "adr_update",
  deviation: "deviation",
  risk_ownership: "risk_ownership",
  lt_escalation: "lt_escalation",
  standard: "standard",
} as const;

export interface CreateReviewOutcome {
  requestId: number;
  sessionId?: number | null;
  decision: CreateReviewOutcomeDecision;
  outcomeType: CreateReviewOutcomeOutcomeType;
  adrReference?: string | null;
  exceptionOwner?: string | null;
  remediationPlan?: string | null;
  riskOwner?: string | null;
  escalationReason?: string | null;
  nextSteps?: string | null;
  notes?: string | null;
  createdBy: string;
}