# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Architecture Review Process Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List all ARRs
      responses:
        "200":
          description: List of architecture review requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArchitectureRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Submit a new ARR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArchitectureRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureRequest"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a specific ARR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Architecture review request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureRequest"
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update ARR status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArchitectureRequest"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureRequest"
...
components:
  schemas:
...
    ArchitectureRequest:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        requestType:
          type: string
          enum: [new_technology, replacement_migration, new_capability, expansion, ma_assessment]
        phase:
          type: string
          enum: [ph1, ph2, ph3]
        submittedBy:
          type: string
        businessUnit:
          type: string
        status:
          type: string
          enum: [submitted, ea_triage, specifications_required, arc_scheduled, arc_review, approved, approved_with_conditions, deferred, rejected]
        priority:
          type: string
          enum: [low, medium, high, critical]
        eaAssignee:
          type: string
          nullable: true
        architectureSpecifications:
          type: string
          nullable: true
        scopeNotes:
          type: string
          nullable: true
        jiraInitiativeId:
          type: integer
          nullable: true
        jiraKey:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - requestType
        - phase
        - submittedBy
        - businessUnit
        - status
        - priority
        - createdAt
        - updatedAt
    CreateArchitectureRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        requestType:
          type: string
          enum: [new_technology, replacement_migration, new_capability, expansion, ma_assessment]
        phase:
          type: string
          enum: [ph1, ph2, ph3]
        submittedBy:
          type: string
        businessUnit:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        architectureSpecifications:
          type: string
          nullable: true
        jiraInitiativeId:
          type: integer
          nullable: true
      required:
        - title
        - description
        - requestType
        - phase
        - submittedBy
        - businessUnit
        - priority
    UpdateArchitectureRequest:
      type: object
      properties:
        status:
          type: string
          enum: [submitted, ea_triage, specifications_required, arc_scheduled, arc_review, approved, approved_with_conditions, deferred, rejected]
        eaAssignee:
          type: string
          nullable: true
        architectureSpecifications:
          type: string
          nullable: true
        scopeNotes:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high, critical]
        jiraInitiativeId:
          type: integer
          nullable: true
...