# lib/api-zod/src/generated/types/architectureRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Architecture Review Process Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ArchitectureRequestPhase } from "./architectureRequestPhase";
import type { ArchitectureRequestPriority } from "./architectureRequestPriority";
import type { ArchitectureRequestRequestType } from "./architectureRequestRequestType";
import type { ArchitectureRequestStatus } from "./architectureRequestStatus";

export interface ArchitectureRequest {
  id: number;
  title: string;
  description: string;
  requestType: ArchitectureRequestRequestType;
  phase: ArchitectureRequestPhase;
  submittedBy: string;
  businessUnit: string;
  status: ArchitectureRequestStatus;
  priority: ArchitectureRequestPriority;
  eaAssignee?: string | null;
  architectureSpecifications?: string | null;
  scopeNotes?: string | null;
  jiraInitiativeId?: number | null;
  jiraKey?: string | null;
  createdAt: Date;
  updatedAt: Date;
}