# lib/api-zod/src/generated/types/createArchitectureRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Architecture Review Process Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateArchitectureRequestPhase } from "./createArchitectureRequestPhase";
import type { CreateArchitectureRequestPriority } from "./createArchitectureRequestPriority";
import type { CreateArchitectureRequestRequestType } from "./createArchitectureRequestRequestType";

export interface CreateArchitectureRequest {
  title: string;
  description: string;
  requestType: CreateArchitectureRequestRequestType;
  phase: CreateArchitectureRequestPhase;
  submittedBy: string;
  businessUnit: string;
  priority: CreateArchitectureRequestPriority;
  architectureSpecifications?: string | null;
  jiraInitiativeId?: number | null;
}