# lib/api-zod/src/generated/types/updateArchitectureRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Architecture Review Process Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateArchitectureRequestPriority } from "./updateArchitectureRequestPriority";
import type { UpdateArchitectureRequestStatus } from "./updateArchitectureRequestStatus";

export interface UpdateArchitectureRequest {
  status?: UpdateArchitectureRequestStatus;
  eaAssignee?: string | null;
  architectureSpecifications?: string | null;
  scopeNotes?: string | null;
  priority?: UpdateArchitectureRequestPriority;
  jiraInitiativeId?: number | null;
}